+++ examples/pip_install/WORKSPACE
@@ -12,7 +12,7 @@ load("@rules_python//python:repositories.bzl", "py_repositories")
 
 py_repositories()
 
-load("@rules_python//experimental/rules_python_external:defs.bzl", "pip_install")
+load("@rules_python//python:pip.bzl", "pip_install")
 
 pip_install(
     # (Optional) You can provide extra parameters to pip.
