--- examples/rules_python_external/README.md
@@ -1,7 +1,0 @@
-# rules_python_external example
-
-This example shows how to use rules_python_external which was developed as an alternative
-to packaging rules in rules_python.
-
-We will soon migrate this example to be the canonical one for rules_python.
-See https://github.com/bazelbuild/rules_python/pull/352 for the longer plan.
