--- experimental/rules_python_external/BUILD
@@ -1,8 +1,0 @@
-filegroup(
-    name = "distribution",
-    srcs = glob(["*.bzl"]) + [
-        "BUILD",
-        "//experimental/rules_python_external/extract_wheels:distribution",
-    ],
-    visibility = ["//:__pkg__"],
-)
