--- experimental/rules_python_external/extract_wheels/__main__.py
@@ -1,5 +1,0 @@
-"""Main entry point."""
-from experimental.rules_python_external.extract_wheels import main
-
-if __name__ == "__main__":
-    main()
