+++ python/pip_install/extract_wheels/__main__.py
@@ -1,0 +1,5 @@
+"""Main entry point."""
+from python.pip_install.extract_wheels import main
+
+if __name__ == "__main__":
+    main()
