+++ python/pip_install/extract_wheels/lib/purelib.py
@@ -2,7 +2,7 @@
 import pathlib
 import shutil
 
-from experimental.rules_python_external.extract_wheels.lib import wheel
+from python.pip_install.extract_wheels.lib import wheel
 
 
 def spread_purelib_into_root(wheel_dir: str) -> None:
