+++ python/pip_install/repositories.bzl
@@ -47,9 +47,11 @@ all_requirements = [name for (name, _, _) in _RULE_DEPS]
 def requirement(pkg):
     return "@pypi__" + pkg + "//:lib"
 
-def rules_python_external_dependencies():
+def pip_install_dependencies():
     """
-    Fetch dependencies these rules depend on. Workspaces that use the rules_python_external should call this.
+    Fetch dependencies these rules depend on. Workspaces that use the pip_install rule can call this.
+    
+    (However we call it from pip_install, making it optional for users to do so.)
     """
     for (name, url, sha256) in _RULE_DEPS:
         maybe(
